Use blank host when running locally
This:
1. Makes it clearer which ones you run locally
2. Stops really wide ipv6 addresses from messing up layout
3. Removes potential unreliable call to external service

I was considering using "localhost", but then that'd be a bit weird
for runs you do by calling it manually on a remote machine, or
when running on AI Platform or whatever.

There is potential improvement by detecting whether this is a
manual run on a remote machine, or AI Platform, or whatever,
but this seems like a step in the right direction.