Print when creating exps & checkpoints
This is slow and user might think training or Replicate is broken.

Ideally this would be a progress bar, but that is a larger piece of
work that will need improvements to storage API. This is an
imperfect stopgap for that.

I considered putting full path in log message, but then it is way
too long and hard to read (includes full exp/chkpt ID!).

Closes #155