Add --json to replicate show
Refactored show to use an options struct rather than accepting
cobra.Command, which means it's possible to test (as per our
conversation about not using integration tests) and makes fetching
flags a bit neater. It would probably make sense to refactor other
commands to do it this way, when we get to them.

Closes #169