Save repository version at project level
This allows us to upgrade to better storage mechanisms in the
future. A project must use a single consistent repository version.

A file called repository.json is written to the top level of the
project's repository. That file is checked on replicate.init(), and if
a newer version is detected, the experiment fails to start.

Signed-off-by: Andreas Jansson <andreas@replicate.ai>